+++ src/mongocxx/options/gridfs/bucket.cpp
@@ -36,12 +36,12 @@ const stdx::optional<std::string>& bucket::bucket_name() const {
     return _bucket_name;
 }
 
-bucket& bucket::chunk_size_bytes(std::size_t chunk_size_bytes) {
+bucket& bucket::chunk_size_bytes(std::int32_t chunk_size_bytes) {
     _chunk_size_bytes = chunk_size_bytes;
     return *this;
 }
 
-const stdx::optional<std::size_t>& bucket::chunk_size_bytes() const {
+const stdx::optional<std::int32_t>& bucket::chunk_size_bytes() const {
     return _chunk_size_bytes;
 }
